restructuring of helper/util functions
This commit moves many parsing/preparation functions from sc.py to other files

- helpers.py is now only containing functions that are useful for the user
- tools.py contains the parsing/preparation function used by sc.py
