+++ sc3nb/helpers.py
@@ -1,82 +1,5 @@
-"""Collection of small helper functions"""
-
-import numbers
-import os
-import re
-import sys
-
+"""Collection of helper functions for the user"""
 import numpy as np
-
-
-def remove_comments(string):
-    """Removes all //single-line or /* multi-line */ c-style comments
-
-    Arguments:
-        string {str} -- Code
-
-    Returns:
-        str -- Code without comments
-    """
-
-    # function code by Onur Yildirim
-    # first group captures quoted strings (double or single)
-    # second group captures comments (//single-line or /* multi-line */)
-    # alternative cares for escaped quotes
-    # pattern = r"(\".*?(?<!\\)\"|\'.*?(?<!\\)\')|(/\*.*?\*/|//[^\r\n]*$)"
-    pattern = r"(\".*?\"|\'.*?\')|(/\*.*?\*/|//[^\r\n]*$)"
-    regex = re.compile(pattern, re.MULTILINE | re.DOTALL)
-
-    def _replacer(match):
-        # if the 2nd group (capturing comments) is not None,
-        # it means we have captured a non-quoted (real) comment string.
-        if match.group(2) is not None:
-            return ""  # so we will return empty to remove the comment
-        else:  # otherwise, we will return the 1st group
-            return match.group(1)  # captured quoted-string
-    return regex.sub(_replacer, string)
-
-def find_executable(executable, path=''):
-    """Looks for executable in os $PATH or specified path
-
-    Arguments:
-        executable {str} -- Executable to be found
-
-    Keyword Arguments:
-        path {str} -- Path at which to look for
-                      executable (default: {''})
-
-    Raises:
-        FileNotFoundError -- Raised if executable
-                             cannot be found
-
-    Returns:
-        str -- Full path to executable
-    """
-
-    if not path:
-        path = os.environ['PATH']
-    paths = path.split(os.pathsep)
-    extlist = ['']
-    if os.name == 'os2':
-        _, ext = os.path.splitext(executable)
-        # executable files on OS/2 can have an arbitrary extension, but
-        # .exe is automatically appended if no dot is present in the name
-        if not ext:
-            executable = executable + ".exe"
-    elif sys.platform == 'win32':
-        pathext = os.environ['PATHEXT'].lower().split(os.pathsep)
-        _, ext = os.path.splitext(executable)
-        if ext.lower() not in pathext:
-            extlist = pathext
-    for ext in extlist:
-        execname = executable + ext
-        if os.path.isfile(execname):
-            return execname
-        for path in paths:
-            file = os.path.join(path, execname)
-            if os.path.isfile(file):
-                return file
-    raise FileNotFoundError("Unable to find sclang executable")
 
 
 def linlin(x, smi, sma, dmi, dma):
